# backend/internal/httpapi/openapi.yaml
openapi: 3.0.3
info:
  title: SpendSense API
  version: "1.0.0"
  description: API for SpendSense Phase 1 (expenses, incomes, wallets, categories, auth)
servers:
  - url: http://localhost:8080
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
  /auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Auth tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
  /auth/refresh:
    post:
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
  /auth/logout:
    post:
      summary: Logout (revoke single refresh token)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: OK
  /auth/logout-all:
    post:
      summary: Logout from all sessions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /ops/refresh-tokens/cleanup:
    post:
      summary: Trigger refresh token cleanup
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cleanup result
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
  /api/v1/currencies:
    get:
      summary: List currencies
      parameters:
        - in: query
          name: default_currency
          schema:
            type: string
      responses:
        '200':
          description: Currency list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrencyListResponse'
  /api/v1/currencies/convert:
    get:
      summary: Convert currency amount
      parameters:
        - in: query
          name: amount
          required: true
          schema:
            type: number
        - in: query
          name: from
          required: true
          schema:
            type: string
        - in: query
          name: to
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Conversion result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrencyConvertResponse'
  /api/v1/dashboard/widgets:
    get:
      summary: Dashboard widgets
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: default_currency
          schema:
            type: string
      responses:
        '200':
          description: Dashboard widgets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardWidgetsResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RefreshRequest:
      type: object
      required: [email, refresh_token]
      properties:
        email:
          type: string
        refresh_token:
          type: string
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        expires_in:
          type: integer
        refresh_token:
          type: string
    LogoutRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
...
    CurrencyOption:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        symbol:
          type: string
        symbol_native:
          type: string
        decimal_digits:
          type: integer
        rounding:
          type: number
        name_plural:
          type: string
        is_default:
          type: boolean
    CurrencyListResponse:
      type: object
      properties:
        default_currency:
          type: string
        currencies:
          type: array
          items:
            $ref: '#/components/schemas/CurrencyOption'
    CurrencyConvertResponse:
      type: object
      properties:
        from_currency:
          type: string
        to_currency:
          type: string
        amount:
          type: number
        converted_amount:
          type: number
        exchange_rate:
          type: number
    DashboardBudgetRow:
      type: object
      properties:
        id:
          type: string
        category_id:
          type: string
        category_name:
          type: string
        category_icon:
          type: string
          nullable: true
        category_color:
          type: string
          nullable: true
        limit:
          type: number
          format: double
        spent:
          type: number
          format: double
        usage_percent:
          type: number
          format: double
        period:
          type: string
        currency:
          type: string
    DashboardMonthlyCashFlowPoint:
      type: object
      properties:
        month:
          type: string
        label:
          type: string
        income:
          type: number
          format: double
        expenses:
          type: number
          format: double
    DashboardWidgetsResponse:
      type: object
      properties:
        base_currency:
          type: string
        budgets:
          type: array
          items:
            $ref: '#/components/schemas/DashboardBudgetRow'
        monthly_cash_flow:
          type: array
          items:
            $ref: '#/components/schemas/DashboardMonthlyCashFlowPoint'