# backend/internal/httpapi/openapi/types.gen.go
// Package openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package openapi

import (
	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Defines values for CreateWalletRequestWalletType.
const (
	BANK         CreateWalletRequestWalletType = "BANK"
	CARD         CreateWalletRequestWalletType = "CARD"
	CASH         CreateWalletRequestWalletType = "CASH"
	MOBILEWALLET CreateWalletRequestWalletType = "MOBILE_WALLET"
)

// CategoryRequest defines model for CategoryRequest.
type CategoryRequest struct {
	IsSystem *bool  `json:"is_system,omitempty"`
	Name     string `json:"name"`
}

// CreateExpenseRequest defines model for CreateExpenseRequest.
type CreateExpenseRequest struct {
	// Amount Decimal amount with 2-digit precision.
	Amount     float64            `json:"amount"`
	CategoryId *int               `json:"category_id,omitempty"`
	Currency   string             `json:"currency"`
	Date       openapi_types.Date `json:"date"`
	Merchant   *string            `json:"merchant,omitempty"`
	Notes      *string            `json:"notes,omitempty"`
	WalletId   int                `json:"wallet_id"`
}

// CreateIncomeRequest defines model for CreateIncomeRequest.
type CreateIncomeRequest struct {
	// Amount Decimal amount with 2-digit precision.
	Amount     float64            `json:"amount"`
	CategoryId *int               `json:"category_id"`
	Currency   string             `json:"currency"`
	IncomeDate openapi_types.Date `json:"income_date"`
	Notes      *string            `json:"notes,omitempty"`
	SourceName string             `json:"source_name"`
	WalletId   int                `json:"wallet_id"`
}

// CreateWalletRequest defines model for CreateWalletRequest.
type CreateWalletRequest struct {
	Currency       string                        `json:"currency"`
	Name           string                        `json:"name"`
	OpeningBalance float64                       `json:"opening_balance"`
	Provider       *string                       `json:"provider"`
	WalletType     CreateWalletRequestWalletType `json:"wallet_type"`
}

// CreateWalletRequestWalletType defines model for CreateWalletRequest.WalletType.
type CreateWalletRequestWalletType string

// CreateWalletTransferRequest defines model for CreateWalletTransferRequest.
type CreateWalletTransferRequest struct {
	// Amount Source wallet amount with 2-digit precision.
	Amount   float64 `json:"amount"`
	Currency string  `json:"currency"`

	// ExchangeRate Required when source and destination wallets use different currencies.
	ExchangeRate *float64           `json:"exchange_rate,omitempty"`
	FeeAmount    *float64           `json:"fee_amount,omitempty"`
	FromWalletId string             `json:"from_wallet_id"`
	Notes        *string            `json:"notes,omitempty"`
	ToWalletId   string             `json:"to_wallet_id"`
	TransferDate openapi_types.Date `json:"transfer_date"`
}

// CurrencyConvertResponse defines model for CurrencyConvertResponse.
type CurrencyConvertResponse struct {
	Amount          *float32 `json:"amount,omitempty"`
	ConvertedAmount *float32 `json:"converted_amount,omitempty"`
	ExchangeRate    *float32 `json:"exchange_rate,omitempty"`
	FromCurrency    *string  `json:"from_currency,omitempty"`
	ToCurrency      *string  `json:"to_currency,omitempty"`
}

// CurrencyListResponse defines model for CurrencyListResponse.
type CurrencyListResponse struct {
	Currencies      *[]CurrencyOption `json:"currencies,omitempty"`
	DefaultCurrency *string           `json:"default_currency,omitempty"`
}

// CurrencyOption defines model for CurrencyOption.
type CurrencyOption struct {
	Code          *string  `json:"code,omitempty"`
	DecimalDigits *int     `json:"decimal_digits,omitempty"`
	IsDefault     *bool    `json:"is_default,omitempty"`
	Name          *string  `json:"name,omitempty"`
	NamePlural    *string  `json:"name_plural,omitempty"`
	Rounding      *float32 `json:"rounding,omitempty"`
	Symbol        *string  `json:"symbol,omitempty"`
	SymbolNative  *string  `json:"symbol_native,omitempty"`
}

// Health defines model for Health.
type Health struct {
	Status *string `json:"status,omitempty"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	Email    string `json:"email"`
	Password string `json:"password"`
}

// LogoutRequest defines model for LogoutRequest.
type LogoutRequest struct {
	RefreshToken string `json:"refresh_token"`
}

// RefreshRequest defines model for RefreshRequest.
type RefreshRequest struct {
	Email        string `json:"email"`
	RefreshToken string `json:"refresh_token"`
}

// RegisterRequest defines model for RegisterRequest.
type RegisterRequest struct {
	Email    openapi_types.Email `json:"email"`
	Password string              `json:"password"`
}

// TokenResponse defines model for TokenResponse.
type TokenResponse struct {
	AccessToken  *string `json:"access_token,omitempty"`
	ExpiresIn    *int    `json:"expires_in,omitempty"`
	RefreshToken *string `json:"refresh_token,omitempty"`
}

// UpdateExpenseRequest defines model for UpdateExpenseRequest.
type UpdateExpenseRequest struct {
	Merchant *string `json:"merchant,omitempty"`
	Notes    *string `json:"notes,omitempty"`
}

// GetApiV1CurrenciesParams defines parameters for GetApiV1Currencies.
type GetApiV1CurrenciesParams struct {
	DefaultCurrency *string `form:"default_currency,omitempty" json:"default_currency,omitempty"`
}

// GetApiV1CurrenciesConvertParams defines parameters for GetApiV1CurrenciesConvert.
type GetApiV1CurrenciesConvertParams struct {
	Amount float32 `form:"amount" json:"amount"`
	From   string  `form:"from" json:"from"`
	To     string  `form:"to" json:"to"`
}

// GetApiV1ExpensesParams defines parameters for GetApiV1Expenses.
type GetApiV1ExpensesParams struct {
	Limit      *int    `form:"limit,omitempty" json:"limit,omitempty"`
	Pagination *string `form:"pagination,omitempty" json:"pagination,omitempty"`
}

// GetApiV1IncomesParams defines parameters for GetApiV1Incomes.
type GetApiV1IncomesParams struct {
	Limit      *int    `form:"limit,omitempty" json:"limit,omitempty"`
	Pagination *string `form:"pagination,omitempty" json:"pagination,omitempty"`
}

// PostApiV1CategoriesJSONRequestBody defines body for PostApiV1Categories for application/json ContentType.
type PostApiV1CategoriesJSONRequestBody = CategoryRequest

// PostApiV1ExpensesJSONRequestBody defines body for PostApiV1Expenses for application/json ContentType.
type PostApiV1ExpensesJSONRequestBody = CreateExpenseRequest

// PutApiV1ExpensesIdJSONRequestBody defines body for PutApiV1ExpensesId for application/json ContentType.
type PutApiV1ExpensesIdJSONRequestBody = UpdateExpenseRequest

// PostApiV1IncomesJSONRequestBody defines body for PostApiV1Incomes for application/json ContentType.
type PostApiV1IncomesJSONRequestBody = CreateIncomeRequest

// PostApiV1WalletsJSONRequestBody defines body for PostApiV1Wallets for application/json ContentType.
type PostApiV1WalletsJSONRequestBody = CreateWalletRequest

// PostApiV1WalletsTransferJSONRequestBody defines body for PostApiV1WalletsTransfer for application/json ContentType.
type PostApiV1WalletsTransferJSONRequestBody = CreateWalletTransferRequest

// PostAuthLoginJSONRequestBody defines body for PostAuthLogin for application/json ContentType.
type PostAuthLoginJSONRequestBody = LoginRequest

// PostAuthLogoutJSONRequestBody defines body for PostAuthLogout for application/json ContentType.
type PostAuthLogoutJSONRequestBody = LogoutRequest

// PostAuthRefreshJSONRequestBody defines body for PostAuthRefresh for application/json ContentType.
type PostAuthRefreshJSONRequestBody = RefreshRequest

// PostAuthRegisterJSONRequestBody defines body for PostAuthRegister for application/json ContentType.
type PostAuthRegisterJSONRequestBody = RegisterRequest